fix submit tests
